:recycle: Refactor every component to remove React.FC
