+++ src/features/pokemonSlice.ts
@@ -8,8 +8,8 @@ import { statusHandlerReducer, wrapReduxAsyncHandler } from "./utilities";
 import { leftPad } from "../utils/leftPad";
 import { baseImageUrl } from "../api/axios";
 
-const INITIAL_SIZE = 1;
-const PAGINATE_SIZE = 1;
+const INITIAL_SIZE = 3;
+const PAGINATE_SIZE = 3;
 
 export type Pokemon = {
   id: number;
