+++ src/index.css
@@ -42,14 +42,3 @@
 body {
   background-color: #f2f1f0;
 }
-
-.card {
-  box-shadow: 0px 10px 30px -5px rgba(0, 0, 0, 0.3);
-  transition: box-shadow 0.5s;
-  will-change: transform;
-  border: 15px solid white;
-}
-
-.card:hover {
-  box-shadow: 0px 30px 100px -10px rgba(0, 0, 0, 0.4);
-}
